move stub provider back to public
