+++ argo/providers/stub/provider.go
@@ -1,4 +1,4 @@
-package internal
+package stub
 
 import (
 	"context"
